Rename WaterHeater cluster files for SCI migration
