+++ src/app/clusters/water-heater-management-server/app_config_dependent_sources.gni
@@ -13,6 +13,7 @@
 # limitations under the License.
 app_config_dependent_sources = [
   "WaterHeaterManagementTestEventTriggerHandler.h",
-  "water-heater-management-server.cpp",
+  "WaterHeaterManagementCluster.cpp",
+  "WaterHeaterManagementCluster.h",
   "water-heater-management-server.h",
 ]
